pyproject.toml
[project]
name = "SMIS"
version = "0.1.0"
description = "School Management Information System for The Spark Playhouse."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    # Olyv framework
    "django>=6.0b1",
    "django-split-settings>=1.3.2",
    "python-decouple>=3.8",
    # User specific
    "gunicorn>=23.0.0",
]
